+++ code/includes/ServerConfiguration.hpp
@@ -27,6 +27,7 @@ public :
 	const size_t								&getMaxClientBodySize(void) const;
 	const std::map<std::string, Route>			&getRoutes(void) const;
 	const Route									*getOneRoutes(std::string path) const;
+	const std::string							getLocation(std::string loc) const;
 
 private :
 	std::vector<std::string>				serverNames;
@@ -34,6 +35,8 @@ private :
 	/// Associate an error with a page, returned to the client.
 	std::map<unsigned short, std::string>	errorPages;
 	size_t									maxClientBodySize;
+	/// @brief string : std::string location, class Route,
+	/// Associate a location, with the corresponding Route.
 	std::map<std::string, Route>			routes;
 	std::string								root;
